Move to using getFullTeam function
Adjust how we retrieve randomized lists of people
based on their being on the team, and/or in the Durham
office to make it more sustainable to add people
in the future instead of doubling the list as we had
in the past. Also "#moffice" has long been a dated term.

Also, update outdated conf.example.js file

Also, update README to add some notes about deployment